+++ app/src/main/java/com/giraffe/tudeeapp/di/appModule.kt
@@ -7,8 +7,8 @@ import com.giraffe.tudeeapp.domain.service.CategoriesService
 import com.giraffe.tudeeapp.domain.service.SplashService
 import com.giraffe.tudeeapp.domain.service.TasksService
 import com.giraffe.tudeeapp.presentation.home.HomeViewModel
-import com.giraffe.tudeeapp.presentation.home.addedittask.TaskEditorViewModel
-import com.giraffe.tudeeapp.presentation.home.taskdetails.TaskDetailsViewModel
+import com.giraffe.tudeeapp.presentation.shared.addedittask.TaskEditorViewModel
+import com.giraffe.tudeeapp.presentation.shared.taskdetails.TaskDetailsViewModel
 import com.giraffe.tudeeapp.presentation.splash.viewmodel.SplashViewModel
 import org.koin.android.ext.koin.androidApplication
 import org.koin.core.module.dsl.viewModel
