+++ app/src/main/java/com/giraffe/tudeeapp/presentation/shared/addedittask/TaskEditorViewModel.kt
@@ -1,4 +1,4 @@
-package com.giraffe.tudeeapp.presentation.home.addedittask
+package com.giraffe.tudeeapp.presentation.shared.addedittask
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -14,7 +14,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
-import kotlinx.datetime.LocalDateTime
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
