+++ app/src/main/java/com/giraffe/tudeeapp/presentation/shared/taskdetails/TaskDetailsState.kt
@@ -1,4 +1,4 @@
-package com.giraffe.tudeeapp.presentation.home.taskdetails
+package com.giraffe.tudeeapp.presentation.shared.taskdetails
 
 import com.giraffe.tudeeapp.presentation.uimodel.TaskUi
